refactor(ui): improve prompt playground robustness and modularity
- Add missing block keywords (else, this, lookup, log) to extractVariables.ts to prevent them from being detected as template variables.
- Split PromptPlaygroundPage.tsx monolithic component into usePromptPlayground hook and separate form/render result subcomponents.
- Remove redundant and stale useEffect in PromptTemplateEditor.tsx.
- Fix unstable useGroupsService dependency in useCallback by wrapping it with useRef.
- Move inline style for empty variables into PromptPlaygroundPage.css.
- Fix React key anti-pattern when rendering validation errors.
- Add additional test cases for extractVariables (now 16 tests).
